Fix clean checkout test resolution
